Extract string operator helpers
